Додав get і put для users/[id]
get і put перевірив, працюють. delete закоментований - бо разом з юзером має видалятись вся супутня інформація по планам і задачам, ще не розібрався
