+++ src/app/api/users/route.ts
@@ -119,43 +119,50 @@ export async function GET() {
   }
 }
 
+export async function PATCH() {
+  return new NextResponse(`Not implemented error`, { status: 501 });
+}
 
-export const DELETE = async (req: Request) => {
-  try {
-    // Видалення всіх користувачів, а також всіх пов'язаних записів у таблицях, таких як 'Plan', 'UserTask' і т.д.
-    const deletedUsers = await prisma.user.deleteMany({
-      where: {},
-    });
+export async function DELETE() {
+  return new NextResponse(`Not implemented error`, { status: 501 });
+}
 
-    // Видалення всіх планів, пов'язаних з користувачами
-    await prisma.plan.deleteMany({
-      where: {},
-    });
+// export const DELETE = async (req: Request) => {
+//   try {
+//     // Видалення всіх користувачів, а також всіх пов'язаних записів у таблицях, таких як 'Plan', 'UserTask' і т.д.
+//     const deletedUsers = await prisma.user.deleteMany({
+//       where: {},
+//     });
 
-    // Ви також можете видаляти інші пов'язані таблиці, як UserTask, PlanTask і т.д., якщо потрібно.
-    await prisma.userTask.deleteMany({
-      where: {},
-    });
+//     // Видалення всіх планів, пов'язаних з користувачами
+//     await prisma.plan.deleteMany({
+//       where: {},
+//     });
 
-    await prisma.planTask.deleteMany({
-      where: {},
-    });
+//     // Ви також можете видаляти інші пов'язані таблиці, як UserTask, PlanTask і т.д., якщо потрібно.
+//     await prisma.userTask.deleteMany({
+//       where: {},
+//     });
 
-    // Повертаємо успішну відповідь
-    return new NextResponse(
-      JSON.stringify({
-        message: "All users and related data deleted successfully",
-        count: deletedUsers.count,
-      }),
-      { status: 200 }
-    );
-  } catch (error: any) {
-    console.error("Error in DELETE request:", error);
-    return new NextResponse(
-      JSON.stringify({ message: "Failed to delete users and related data", error: error.message }),
-      { status: 500 }
-    );
-  }
-};
+//     await prisma.planTask.deleteMany({
+//       where: {},
+//     });
+
+//     // Повертаємо успішну відповідь
+//     return new NextResponse(
+//       JSON.stringify({
+//         message: "All users and related data deleted successfully",
+//         count: deletedUsers.count,
+//       }),
+//       { status: 200 }
+//     );
+//   } catch (error: any) {
+//     console.error("Error in DELETE request:", error);
+//     return new NextResponse(
+//       JSON.stringify({ message: "Failed to delete users and related data", error: error.message }),
+//       { status: 500 }
+//     );
+//   }
+// };
